move test to proper file
